+++ src/nav/nav-bar.js
@@ -1,36 +1,58 @@
 import React from "react"
-import { Dropdown, Menu, Responsive } from "semantic-ui-react"
-import { aboutRoute, connectRoute, sermonsRoute, sundayRoute } from "../common/url-helper"
+import styled from "styled-components"
+import Theme from "../common/theme"
+import { Link } from "react-router-dom"
+import { aboutRoute,connectRoute,sermonsRoute,sundayRoute } from "../common/url-helper"
 
-const NavBar = props => {
-  const { onClick } = props
+const ClearFix = styled.div`
+  clear: both;
+`
 
+const Logo = styled.img`
+  float: left;
+  height: 40px;
+  width: 40px;
+`
+
+const NavLinks = styled.ul`
+  a {
+    color: ${Theme.colors.light.primary}
+  }
+
+  a:hover {
+    color: ${Theme.colors.light.subtle}
+  }
+
+  .dark-mode & {
+    a { 
+      color: ${Theme.colors.dark.primary}
+    }
+
+    a:hover {
+      color: ${Theme.colors.dark.subtle}
+    }
+  }
+`
+const NavItem = styled.li`
+  display: inline;
+  float: right;
+  line-height: ${Theme.lineHeight.xl};
+  margin-left: ${Theme.spacing.sm};
+`
+
+const NavBar = ({ darkMode }) => {
+  const className = darkMode ? "dark-mode" : null
   return (
-    <Menu secondary>
-      <Menu.Item>
-        <img src="/cma.png" style={{ width: 40, height: 40 }} />
-      </Menu.Item>
-
-      <Responsive as={Menu.Menu} position="right" maxWidth={767}>
-        <Menu.Item>
-          <Dropdown icon="chevron down">
-            <Dropdown.Menu>
-              <Dropdown.Item text="Connect" name="connect" onClick={onClick} />
-              <Dropdown.Item text="Sermons" name="sermons" onClick={onClick} />
-              <Dropdown.Item text="About" name="about" onClick={onClick} />
-              <Dropdown.Item text="Sunday" name="sunday" onClick={onClick} />
-            </Dropdown.Menu>
-          </Dropdown>
-        </Menu.Item>
-      </Responsive>
-
-      <Responsive as={Menu.Menu} position="right" minWidth={768}>
-        <Menu.Item name="connect" onClick={onClick} />
-        <Menu.Item name="sermons" onClick={onClick} />
-        <Menu.Item name="about" onClick={onClick} />
-        <Menu.Item name="sunday" onClick={onClick} />
-      </Responsive>
-    </Menu>
+    <nav className={className}>
+      <Logo src="/cma.png" />
+      <NavLinks>
+        <NavItem><Link to={sundayRoute.url()}>Sunday</Link></NavItem>
+        <NavItem><Link to={aboutRoute.url()}>About</Link></NavItem>
+        <NavItem><Link to={sermonsRoute.url()}>Sermons</Link></NavItem>
+        <NavItem><Link to={connectRoute.url()}>Connect</Link></NavItem>
+      </NavLinks>
+      <ClearFix />
+    </nav>
   )
 }
