+++ src/stories/index.js
@@ -1,6 +1,6 @@
 import "semantic-ui-css/semantic.min.css"
 import Footer from "../common/footer"
-import NavBar from "../nav/nav-bar_v2"
+import NavBar from "../nav/nav-bar"
 import React from 'react';
 import { BrowserRouter } from "react-router-dom"
 import { action } from '@storybook/addon-actions';
